requirements.txt
Django==1.7.1
PyJWT==0.3.0
PyPDF2==1.23
defusedxml==0.4.1
mysqlclient==1.3.4
numpy==1.9.1
oauthlib==0.7.1
python-social-auth==0.2.1
python3-openid==3.0.5
requests==2.4.3
requests-oauthlib==0.3.1
six==1.8.0
djangorestframework
markdown
django-filter
